Separando Publication list em um novo arquivo
