+++ src/components/Publication.js
@@ -12,9 +12,8 @@ import { editPostRequest, likeRequest, dislikeRequest } from "../services/api";
 import ApplicationContext from "../contexts/ApplicationContext";
 
 import DeleteModal from "./DeleteModal"
-import { FiRefreshCw } from "react-icons/fi";
 
-function Publication({
+export default function Publication({
     userLiked,
     firstLike,
     secondLike,
@@ -226,73 +225,6 @@ function Publication({
     );
 };
 
-export default function PublicationList({ posts, numberVisiblePosts, setNumberVisiblePosts }) {
-    function checkForPosts() {
-        if (!posts) {
-            return (
-                <TextContainer>
-                    <h2>Loading...</h2>
-                </TextContainer>
-            );
-        } else if (numberVisiblePosts === 0) {
-            return (
-                <TextContainer>
-                    <h2>There are no posts yet</h2>
-                </TextContainer>
-            );
-        } else {
-            return (
-                posts.slice(posts.length - numberVisiblePosts).map(item => (
-                    <Publication
-                        key={Date.now() * item.id}
-                        userLiked={item.userliked}
-                        firstLike={item.firstlike}
-                        secondLike={item.secondlike}
-                        likesAmount={item.likes_amount}
-                        postId={item.id}
-                        userImage={item.pic_url}
-                        userName={item.name}
-                        authorId={item.author_id}
-                        postTitle={item.content}
-                        postLink={item.link_url}
-                        LinkName={item.link_title}
-                        LinkSummary={item.link_description}
-                        LinkImg={item.link_image}
-                        userauthorship={item.userauthorship}
-                    />))
-            );
-        };
-    };
-
-    function renderLoadMoreButton() {
-        if (posts.length - numberVisiblePosts > 1) {
-            return (
-                <LoadMoreButton onClick={() => setNumberVisiblePosts(posts.length)}>
-                    {`${posts.length - numberVisiblePosts} new posts, load more!`}
-                </LoadMoreButton>
-            )
-        }
-        else if (posts.length - numberVisiblePosts > 0) {
-            return (
-                <LoadMoreButton onClick={() => setNumberVisiblePosts(posts.length)}>
-                    <span>{`${posts.length - numberVisiblePosts} new post, load more!`}</span>
-                    <FiRefreshCw />
-                </LoadMoreButton>
-            )
-        }
-        return ""
-    }
-
-    const renderPosts = checkForPosts();
-    const loadMoreButton = renderLoadMoreButton();
-
-    return (
-        <>
-            {loadMoreButton}
-            {renderPosts}
-        </>
-    )
-};
 
 const Post = styled.div`
     background: #171717;
@@ -440,35 +372,5 @@ const ContentInput = styled.input`
     :disabled {
         background-color: lightgray;
         color: darkgray;
-    }
-`;
-
-const LoadMoreButton = styled.button`
-    width:100%;
-    padding: 16px 18px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #1877F2;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    border-radius: 16px;
-    border: none;
-    font-weight: 400;
-    font-size: 16px;
-    font-family: 'Lato';
-    color: #FFFFFF;
-
-    span{
-        margin-right:10px;
-    }
-`;
-
-const TextContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    h2{
-        font: 400 20px 'Oswald', sans-serif;
-        color: #FFFFFF;
     }
 `;
